Refactor sum implementation
