+++ cpp/src/arrow/compute/kernels/sum-internal.h
@@ -20,9 +20,12 @@
 #include <memory>
 #include <type_traits>
 
+#include "arrow/compute/kernel.h"
+#include "arrow/compute/kernels/aggregate.h"
 #include "arrow/status.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
+#include "arrow/util/bit-util.h"
 
 namespace arrow {
 
@@ -50,6 +53,153 @@ struct FindAccumulatorType<I, typename std::enable_if<IsUnsignedInt<I>::value>::
 template <typename I>
 struct FindAccumulatorType<I, typename std::enable_if<IsFloatingPoint<I>::value>::type> {
   using Type = DoubleType;
+};
+
+template <typename ArrowType, typename StateType>
+class SumAggregateFunction final : public AggregateFunctionStaticState<StateType> {
+  using CType = typename TypeTraits<ArrowType>::CType;
+  using ArrayType = typename TypeTraits<ArrowType>::ArrayType;
+
+  static constexpr int64_t kTinyThreshold = 32;
+  static_assert(kTinyThreshold > 18,
+                "ConsumeSparse requires at least 18 elements to fit 3 bytes");
+
+ public:
+  Status Consume(const Array& input, StateType* state) const override {
+    const ArrayType& array = static_cast<const ArrayType&>(input);
+
+    if (input.null_count() == 0) {
+      *state = ConsumeDense(array);
+    } else if (input.length() <= kTinyThreshold) {
+      // In order to simplify ConsumeSparse implementation (requires at least 3
+      // bytes of bitmap data), small arrays are handled differently.
+      *state = ConsumeTiny(array);
+    } else {
+      *state = ConsumeSparse(array);
+    }
+
+    return Status::OK();
+  }
+
+  Status Merge(const StateType& src, StateType* dst) const override {
+    *dst += src;
+    return Status::OK();
+  }
+
+  Status Finalize(const StateType& src, Datum* output) const override {
+    *output = src.Finalize();
+    return Status::OK();
+  }
+
+  std::shared_ptr<DataType> out_type() const override { return StateType::out_type(); }
+
+ private:
+  StateType ConsumeDense(const ArrayType& array) const {
+    StateType local;
+
+    const auto values = array.raw_values();
+    const int64_t length = array.length();
+    for (int64_t i = 0; i < length; i++) {
+      local.sum += values[i];
+    }
+
+    local.count = length;
+
+    return local;
+  }
+
+  StateType ConsumeTiny(const ArrayType& array) const {
+    StateType local;
+
+    internal::BitmapReader reader(array.null_bitmap_data(), array.offset(),
+                                  array.length());
+    const auto values = array.raw_values();
+    for (int64_t i = 0; i < array.length(); i++) {
+      if (reader.IsSet()) {
+        local.sum += values[i];
+        local.count++;
+      }
+      reader.Next();
+    }
+
+    return local;
+  }
+
+  inline StateType UnrolledSum(uint8_t bits, const CType* values) const {
+    StateType local;
+
+    if (bits < 0xFF) {
+#define SUM_SHIFT(ITEM) values[ITEM] * static_cast<CType>(((bits >> ITEM) & 1U))
+      // Some nulls
+      local.sum += SUM_SHIFT(0);
+      local.sum += SUM_SHIFT(1);
+      local.sum += SUM_SHIFT(2);
+      local.sum += SUM_SHIFT(3);
+      local.sum += SUM_SHIFT(4);
+      local.sum += SUM_SHIFT(5);
+      local.sum += SUM_SHIFT(6);
+      local.sum += SUM_SHIFT(7);
+      local.count += BitUtil::kBytePopcount[bits];
+#undef SUM_SHIFT
+    } else {
+      // No nulls
+      for (size_t i = 0; i < 8; i++) {
+        local.sum += values[i];
+      }
+      local.count += 8;
+    }
+
+    return local;
+  }
+
+  StateType ConsumeSparse(const ArrayType& array) const {
+    StateType local;
+
+    // Sliced bitmaps on non-byte positions induce problem with the branchless
+    // unrolled technique. Thus extra padding is added on both left and right
+    // side of the slice such that both ends are byte-aligned. The first and
+    // last bitmap are properly masked to ignore extra values induced by
+    // padding.
+    //
+    // The execution is divided in 3 sections.
+    //
+    // 1. Compute the sum of the first masked byte.
+    // 2. Compute the sum of the middle bytes
+    // 3. Compute the sum of the last masked byte.
+
+    const int64_t length = array.length();
+    const int64_t offset = array.offset();
+
+    // The number of bytes covering the range, this includes partial bytes.
+    // This number bounded by `<= (length / 8) + 2`, e.g. a possible extra byte
+    // on the left, and on the right.
+    const int64_t covering_bytes = BitUtil::CoveringBytes(offset, length);
+
+    // Align values to the first batch of 8 elements. Note that raw_values() is
+    // already adjusted with the offset, thus we rewind a little to align to
+    // the closest 8-batch offset.
+    const auto values = array.raw_values() - (offset % 8);
+
+    // Align bitmap at the first consumable byte.
+    const auto bitmap = array.null_bitmap_data() + BitUtil::RoundDown(offset, 8) / 8;
+
+    // Consume the first (potentially partial) byte.
+    const uint8_t first_mask = BitUtil::kTrailingBitmask[offset % 8];
+    local += UnrolledSum(bitmap[0] & first_mask, values);
+
+    // Consume the (full) middle bytes. The loop iterates in unit of
+    // batches of 8 values and 1 byte of bitmap.
+    for (int64_t i = 1; i < covering_bytes - 1; i++) {
+      local += UnrolledSum(bitmap[i], &values[i * 8]);
+    }
+
+    // Consume the last (potentially partial) byte.
+    const int64_t last_idx = covering_bytes - 1;
+    const uint8_t last_mask = BitUtil::kPrecedingWrappingBitmask[(offset + length) % 8];
+    local += UnrolledSum(bitmap[last_idx] & last_mask, &values[last_idx * 8]);
+
+    return local;
+  }
 };
 
 }  // namespace compute
